requirements.txt
aws-cdk-lib==2.96.2
constructs>=10.0.0,<11.0.0
cdk-aws-iotfleetwise>=0.2.14
